backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<3.16
gunicorn>=20.0,<22.0
google-cloud-storage>=2.0,<3.0
whitenoise>=6.0,<7.0 # For serving static files
python-dotenv>=1.0.0,<2.0.0 # For loading .env files
mssql-django>=1.0,<2.0 # For MSSQL database connection
django-cors-headers>=3.0,<4.0 # For CORS handling
google-genai==1.7.0
google-auth>=2.26.0